Adjusted the logic processing of noPay
