+++ lowcode/scattertrigger.go
@@ -66,13 +66,13 @@ func (scatterTriggerData *ScatterTriggerData) BuildPBComponentData() proto.Messa
 
 // GetVal -
 func (scatterTriggerData *ScatterTriggerData) GetVal(key string) int {
-	if key == STDVSymbolNum {
+	if key == CVSymbolNum {
 		return scatterTriggerData.SymbolNum
-	} else if key == STDVWildNum {
+	} else if key == CVWildNum {
 		return scatterTriggerData.WildNum
-	} else if key == STDVRespinNum {
+	} else if key == CVRespinNum {
 		return scatterTriggerData.RespinNum
-	} else if key == STDVWins {
+	} else if key == CVWins {
 		return scatterTriggerData.Wins
 	}
 
@@ -81,13 +81,13 @@ func (scatterTriggerData *ScatterTriggerData) GetVal(key string) int {
 
 // SetVal -
 func (scatterTriggerData *ScatterTriggerData) SetVal(key string, val int) {
-	if key == STDVSymbolNum {
+	if key == CVSymbolNum {
 		scatterTriggerData.SymbolNum = val
-	} else if key == STDVWildNum {
+	} else if key == CVWildNum {
 		scatterTriggerData.WildNum = val
-	} else if key == STDVRespinNum {
+	} else if key == CVRespinNum {
 		scatterTriggerData.RespinNum = val
-	} else if key == STDVWins {
+	} else if key == CVWins {
 		scatterTriggerData.Wins = val
 	}
 }
@@ -253,9 +253,9 @@ func (scatterTrigger *ScatterTrigger) InitEx(cfg any, pool *GamePropertyPool) er
 		scatterTrigger.Config.WinMulti = 1
 	}
 
-	if scatterTrigger.Config.BetType == BTypeNoPay {
-		scatterTrigger.Config.NeedDiscardResults = true
-	}
+	// if scatterTrigger.Config.BetType == BTypeNoPay {
+	// 	scatterTrigger.Config.NeedDiscardResults = true
+	// }
 
 	scatterTrigger.onInit(&scatterTrigger.Config.BasicComponentConfig)
 
@@ -316,12 +316,12 @@ func (scatterTrigger *ScatterTrigger) canTrigger(gameProp *GameProperty, gs *sgc
 				}, false)
 
 			if ret != nil {
-				if scatterTrigger.Config.BetType == BTypeNoPay {
-					ret.CoinWin = 0
-					ret.CashWin = 0
-				} else {
-					// gameProp.ProcMulti(ret)
-				}
+				// if scatterTrigger.Config.BetType == BTypeNoPay {
+				// 	ret.CoinWin = 0
+				// 	ret.CashWin = 0
+				// } else {
+				// 	// gameProp.ProcMulti(ret)
+				// }
 
 				isTrigger = true
 
@@ -334,10 +334,10 @@ func (scatterTrigger *ScatterTrigger) canTrigger(gameProp *GameProperty, gs *sgc
 		})
 
 		if ret != nil {
-			if scatterTrigger.Config.BetType == BTypeNoPay {
-				ret.CoinWin = 0
-				ret.CashWin = 0
-			} else {
+			if scatterTrigger.Config.BetType != BTypeNoPay {
+				// 	ret.CoinWin = 0
+				// 	ret.CashWin = 0
+				// } else {
 				if scatterTrigger.Config.SymbolCodeCountScatterPayAs > 0 {
 					ret.Mul = gameProp.CurPaytables.MapPay[scatterTrigger.Config.SymbolCodeCountScatterPayAs][ret.SymbolNums-1]
 					ret.CoinWin = gameProp.CurPaytables.MapPay[scatterTrigger.Config.SymbolCodeCountScatterPayAs][ret.SymbolNums-1]
@@ -361,10 +361,10 @@ func (scatterTrigger *ScatterTrigger) canTrigger(gameProp *GameProperty, gs *sgc
 			})
 
 		if ret != nil {
-			if scatterTrigger.Config.BetType == BTypeNoPay {
-				ret.CoinWin = 0
-				ret.CashWin = 0
-			} else {
+			if scatterTrigger.Config.BetType != BTypeNoPay {
+				// 	ret.CoinWin = 0
+				// 	ret.CashWin = 0
+				// } else {
 				if scatterTrigger.Config.SymbolCodeCountScatterPayAs > 0 {
 					ret.Mul = gameProp.CurPaytables.MapPay[scatterTrigger.Config.SymbolCodeCountScatterPayAs][ret.SymbolNums-1]
 					ret.CoinWin = gameProp.CurPaytables.MapPay[scatterTrigger.Config.SymbolCodeCountScatterPayAs][ret.SymbolNums-1]
@@ -394,6 +394,15 @@ func (scatterTrigger *ScatterTrigger) canTrigger(gameProp *GameProperty, gs *sgc
 
 // procWins
 func (scatterTrigger *ScatterTrigger) procWins(std *ScatterTriggerData, lst []*sgc7game.Result) (int, error) {
+	if scatterTrigger.Config.BetType == BTypeNoPay {
+		for _, v := range lst {
+			v.CoinWin = 0
+			v.CashWin = 0
+		}
+
+		return 0, nil
+	}
+
 	std.WinMulti = scatterTrigger.GetWinMulti(&std.BasicComponentData)
 
 	for _, v := range lst {
